requirements.txt
PyYAML
requests
fastapi
langchain
langgraph
uvicorn
twilio
orjson
langchain_groq
langchain_experimental
langchain_openai
redis
python-dotenv
psycopg_pool
psycopg
psycopg[binary]
asyncpg
pydantic
langgraph-checkpoint-redis
python-multipart
boto3
beautifulsoup4
requests
playwright
chromadb
sentence_transformers
langchain_chroma
passlib[bcrypt]
itsdangerous